correction

+++ 0x00-python_variable_annotations/5-sum_list.py
@@ -7,7 +7,7 @@ returns their sum as a float. """
 
 def sum_list(input_list: list([float])) -> float:
     """ takes a list input_list of floats and return their sum"""
-    summ: float = 0.0
+    summ: int = 0
     for i in input_list:
             summ += i
